dotfile variants

+++ derivations/dotfiles.nix
@@ -6,8 +6,8 @@ pkgs:
     # src :: derivation
     src,
 
-    # files :: Map string { src? :: string -> derivation; subs :: Theme -> attrset }
-    files
+    # filesFn :: string -> Theme -> Map string { variant? :: string; subs :: attrset }
+    filesFn
 }:
 
 # string -> Theme -> derivation
@@ -17,31 +17,46 @@ let
     std = pkgs.lib;
 
     inherit (builtins)
+        baseNameOf
         concatStringsSep
+        dirOf
         toString;
 
     inherit (std.attrsets)
         mapAttrsToList;
 
+    inherit (std.strings)
+        optionalString;
+
     inherit (lib)
         sedScript;
-    
+
     commands =
-        mapAttrsToList
-        (file: themedDotfile:
-            let
-                _src =
-                    if themedDotfile ? "src"
-                    then themedDotfile.src
-                    else src + "/${file}";
-                subs = themedDotfile.subs theme; 
-            in ''
-                outfile=$out/${file}
-                mkdir -p $(dirname $outfile)
-                sed ${sedScript subs} ${_src} > $outfile
-            ''
-        )
-        files;
+        let
+            files = filesFn themeName theme;
+        in
+            mapAttrsToList
+            (file: dotfile:
+                let
+                    srcfile =
+                        if dotfile ? "variant"
+                        then
+                            let
+                                dir = dirOf file;
+                            in
+                                src +
+                                (
+                                    (optionalString (dir != ".") "/${dir}") +
+                                    "/${dotfile.variant}.${baseNameOf file}"
+                                )
+                        else src + "/${file}";
+                in ''
+                    outfile=$out/${file}
+                    mkdir -p $(dirname $outfile)
+                    sed ${sedScript dotfile.subs} ${srcfile} > $outfile
+                ''
+            )
+            files;
 in
 pkgs.runCommandLocal "dotfiles"
 {}
@@ -53,7 +68,18 @@ mkdir -p $out
 files=$(find ${src} -type f)
 for srcfile in $files; do
     file=''${srcfile#${src}/}
+
     outfile=$out/$file
+    outdir=$(dirname $outfile)
+
+    basename=$(basename $file)
+    varoutbasename=$(echo $basename | sed "s/^[^.]*\.//")
+    varoutfile=$outdir/$varoutbasename
+
+    if [ -e $varoutfile ]; then
+        continue
+    fi
+
     if [ ! -e $outfile ]; then
         mkdir -p $(dirname $outfile)
         cp $srcfile $outfile
